inline AdcTimeStamp to remove link dependencies.
